Fix bug in device.sync that prevented it from working on some drives
device.sync wasn't working when the source folder was on a different drive than the temp directory.

I think the intention of the code was to get the path without the drive specifier by getting the relative path of it when compared to the drive specifier. However the temp file driver specifier was being passed through rather than the source file one.