Documentation note that sync deletes all other files on device.
I hadn't realized when I ran sync that it would delete all the other files on my device, so I figure it's worth making it clear, especially as a common way of using Thonny is actually to do development directly on the device files, and periodically backup, which is what I had been doing. Luckily I had a recent backup.